move tlist stuff to seprate file
