+++ src/tile.c
@@ -61,63 +61,6 @@ bool tile_parse(FILE* file, tile** t) {
     return false;
 }
 
-bool tlist_parse(const char* filename, sized_tlist* list) {
-    if (!list && !list->list) {
-        return false;
-    }
-    FILE* file;
-    if ((file = fopen(filename, "r")) == 0) {
-        return false;
-    }
-    for (size_t i = 0; i < list->len; ++i) {
-        // if any tile parsing fails return false
-        if (!tile_parse(file, &list->list[i])) {
-            fclose(file);
-            return false;
-        }
-    }
-    fclose(file);
-    return true;
-}
-
-size_t tlist_get_len(const char* filename) {
-    FILE* list = fopen(filename, "r");
-    size_t count = 0;
-    if (list) {
-        int ch;
-        while ((ch = getc(list)) != EOF) {
-            while ((ch = getc(list)) != EOF && isspace(ch)) {
-                continue;
-            }
-            if (ch != EOF && !isspace(ch)) {
-                ++count;
-            }
-            while ((ch = getc(list)) != EOF && !isspace(ch)) {
-                continue;
-            }
-        }
-        fclose(list);
-    }
-    return count;
-}
-
-bool tlist_init(const char* filename, sized_tlist* list) {
-    list->len = tlist_get_len(filename);
-    list->list = malloc(sizeof(tile*) * list->len);
-    memset(list->list, 0, sizeof(tile*) * list->len);
-    return tlist_parse(filename, list);
-}
-
-sized_tlist tlist_init_exit_on_err(const char* filename) {
-    sized_tlist list;
-    if (!tlist_init(filename, &list)) {
-        free(list.list);
-        fputs("error parsing tile list\n", stderr);
-        exit(EXIT_FAILURE);
-    }
-    return list;
-}
-
 char elem_to_char(element e) {
     switch (e) {
     case CASTLE: return 'c';
@@ -157,13 +100,6 @@ char* tile_to_str_alloc(const tile* t) {
 void tile_print(const tile* t) {
     char buff[5];
     printf("%.*s", 5, tile_to_str(t, buff));
-}
-
-void tlist_print(const sized_tlist* list) {
-    for (size_t i = 0; i < list->len; ++i) {
-        tile_print(list->list[i]);
-        putchar('\n');
-    }
 }
 
 tile* tile_rotate(tile* t) {
